sap.c: check if SAP message is a deletion message
When the service user performs the function received_announcement,
defined it sap.c, it now, checks if the SAP header have bit number 5 set
to 1. To do so, it compares byte 1 of SAP header to 0x24 (value of byte
1 when for a deletion message).
When the deletion message is received, the Service user stops to listen
on the socket: it leaves the multicast group.