Fixes a segfault bug in receive_announcement
When the SU receivethe deletion message, it deletes the pipeline it had
created to display the stream. However, when the pipeline was deleted,
it was freeing the sap_data associated to the channel. However, it is
not the behaviour we want. Because otherwise, the SU cannot listen for
other SAP/SDP announcements.