Fix a bug in update_channelEntry
update_channelEntry is a function of channelTable.c used when the user
wishes to change the VideoFormatTable associated to the channel. To do
so we update the stream_data of the channel with the stream_data of the
new videoFormat. However, the multicast IP that was used to build the
pipeline of those stream_data, was the multicast IP computed with the
previous channel Number.
For example if we have two videoFormats ( 2 sources ) during the system
initialisation 2 channels are created (one for each videoFormat). So we
have something like that:
VF1 --> channel1 --> 239.192.1.119
VF2 --> channel2 --> 239.192.2.119

if we want to change that into:
VF1 --> channel1 --> 239.192.1.119
VF1 --> channel2 --> 239.192.2.119

But the stream_data (which contains the pipeline and udpsink element) of
VF1 have been built with the multicast IP 239.192.1.119, when we whange
the VF associated to channel2, we expect that this VF will be streamed
on the multicast IP 239.192.2.119. So here's the problem. What this
commit does, it that after updating the stream_data of the channel2 with
the stream_data of VF1, it set the pipeline of in NULL state and updates
the IP of udpSink element.