sdp.c: Change the way to retireve SDP message
SDP message is retrieved from the SAP message using the function
get_SDP(). Previously, this function, initated the pipeline used to
get the stream from the multicast IP (and decode if needed) and started
to display the stream. Then it returned TRUE on success and FALSE on
failure. Furthermore, on a other hand, to avoid building the pipeline
over and over each time the serviceUser's gets the SAP/SDP annoucement,
I had implemented a system that ignore SAP/SDP messages got on the
channel if that are identical to the previous SAP/SDP message received.
This system worked well if only one videoChannel is started. But if the
sevriceUser gets SAP/SDP messages from several videoChannel, it does not
start the stream.
This commit doesn't fix the problem, but it's a first step to solution.