sap.c: change the way to receive SAP/SDP messages
Before SAP/SDP were received per channel. Indeed, each time a channel
(serviceUser) was set to start, a callback function was added to the
MainLoop. This callback read the socket 224.127.2.254:9875, depay the
SAP message ot get the SDP message, checked the value of the mutlicast
IP contained in it and if it was the same as the one in
channelReceiveIP, it started a pipeline and displayed the stream. If the
stream was already displayed, this callback just ignored the SAP/SDP
announcement. However, this implementation was not right. The following
case did not work:
_ 2 serviceUser that displayed the stream got from 1 videoChannel.

The problem was the following: the read function works as follows: it
take the input file pointed into the kernel by the file descriptor
associated to it, and it puts it into the buffer passed in parameter.
Thus, if two reads are perform consecutively on the same socket, the
second read return -1 (it does not read any data, because those data
have already been read before). In our case, 1 single
socket(224.2.127.254:9875) was read twice (consicutely): once by the
serviceUser 1, a second time by the serviceUser 2. Thus, only the socket
read by serviceUser 1 was disaplyed (launched a steam).

The new solution is that, the same callback is called every second,
however, it is not associateed to a channel in particuler, instead, it
iteratres through the channelTable, compares the channelReceiveIp
parameter of each entries and starts the stream if this match. This
works way better. But it hurts performance, as each seconds it iterates
through all the channelTalbe.