create a second table: channelTable_SU
This stable contains only channel that are
_ ServiceUser
_ channelStatus is start
It means that it contains on the channelTable_entries that we should
iterates when checking if the channelRecevieIp matches with the one in
SDP message receive. This is way more efficient than iterates through
all the channelTable, we only iterates through the started serviceUser.