Add the number if SU channel in configuration file
Before an empty channel was added each time a channel was used. So the
user always had a channel available (a serviceUser) so he can plays a
stream on it. Now the number of channels is predifined into the
configuration file vivoe_mib.conf before launching the program. When the
program is started, it reads into vivoe_mib.conf the value
"nb_screens" , it initializes channelNumber with it, then when
channelTable is initialized in initialize_channelTablei() (called when
openvivoe_deamon is started) it gets the number of empty channel(s) it
needs to create in channelNumber variable.