mib-config.c: remove function check_MAC
This function was used to check if a string as the MAC address format:
XX:XX:XX:XX:XX:XX. It was used in a version of openvivoe in which the
ethernet addresses on ethernet interfaces was entered by the user in the
configuration file. Now, this values are automatically detected, there
is no need to keep this function anymore.