Replace fakesource by a real one
The source was created by the function source_creation. This builds a
GstElement using factory_make() function, this element is a videotestsrc
(the famous mires). But the program aims to work with a real camera. In
this commit videotestsrc is not used anymore, instead we use v4l2src,
setting its "device" parameter to /dev/video0. Quick tests have been
done and it seems to works just fine.