Create source using gst_parse_launch
I tried (for test purposes) to create a source using gst_fatory_make()
function. However, what would be nice is to give to the user the
possibility to give through the configuration file the gst-launch
command to read th video from the source. Then using gst_parse_launch
we can build a pipeline. This is what I did. My first idea was to create
direclty the pipeline using gst_parse_laucnh but the variable was
created locally which makes problems. The solution was to create the
pipeline direclty using gst_parsel_launch however, we need to have
pipelines built using gst_pipeline_new for the service user. So the most
scalable solution is to create a bin using gst_bin_parse_launch, and then
add this bin to the pipeline. This will also be the last element of the
pipeline so no need to iterates onto the elements of the pipeline as it
was done before.