Get the source video from vivoe_mib.conf file
The user should have the possibility to initiate the source from the
configuration file. We want the user to have the possibility to give in
the configuration file a Gstreamer command line that will be used by open
VIVOE as the way to get the input video source. This is implemented by
this commit. Each time a source is launched, it will parse the
configuration file vivoe_mib.conf. It will get the group name :
[source_num]
Where num is the value of videoFormatNumber + 1 (indeed,
the indexes of sources newly created is the number of source already
created, thus vidoeFormatNumber increased by 1). When the group is found,
it will look for the key name "gst_source" which can be easily modified
in mib_conf. This key gets as value a string corresponding to the
gstreamer command line used to get the video source.