stream_registration.c: map colorspace to sampling for J2K
J2k videos have caps with a field named "colorspace", however, this field
corresponds implicitely to the field named sampling in the MIB and in
RAW video caps. This commit add a function in stream_regiustration.c
that mapps the value of colorspace to a usable value of sampling.
After a runnig "gst-inspect-1.0 openjpegenc" we can see that this
plug-in have in src pads only 3 possible colorspace values:
_ sRGB corresponds to sampling: RGB
_ sYUV corresponds to sampling: YCbCr-4:2:2 (for example)
_ GRAY corresponds to sampling: GRAYSCALE

It have been tested with videotestsrc and JPEG200 works just fine.