stop_program: change the way to free data.
Data of channels where deleting in the function delete_stream, only if
they were a videoChannel. However, I do have reasons to think that it
can cause problems with webcams. Indeed, when we exit the program, we
need to quit it properly: it means quitting the pipeline nicely (unref
pipeline). Indeed, I noticed that otherwise it causes problems with a
camera: when the program is stopped and started twice: the second time:
it blocks inside the function start_streaming(), when we try to set the
pipeline to PLAYING state. A proper use of the variable GST_DEBUG=5
gives us the following information: it seems that the second time the
program fails to open /dev/video0, which leads to a timeout on the SNMP
request to set channelStatus to start.
First thing that this commit does, is to catch SIGINT and SIGTERM
properly and delete all the table dynamically allocated
(ethernetIfTable, videoFormatTable, channelTable). But before deleting
the channelTable, we launch the function dele_stream, that should set
pipeline to NULL state (I hope this is the proper way to end a Gstreamer
binary, launch the DEBUG option on a gst-launch should answer this
question).