Add a "maintenance_group" flag to MIB's parameters
The VIVOE norm specifies 3 different modes of operation.
1/ the normal mode which is the first one that have been implemented
2/ defaultStartUp mode, which will be implemented a little later
3/ maintenance mode, which is the object of this commit

The maintenance mode is here to provide extra security to openvivoe. It
is a mode in  which values of critical parameters can be modify. It
means that before modifying those parameters the device should be set in
maintenance mode.The parameters for which this rule applies belong to a
group named maintenance_group. This commit adds two things to the code in
order to implement this functionnality. First it adds a
"maintenance_flag" to MIB's parameters, represented by a boolean member
in the structure parameter of mib_parameter.h. Then it adds in
mib-config.c a function that is used to: set the maintenance_flag of a
parameter.
Furthermore, I notice that not all parameter's intiations are not done
in the same file. I corrected it. Now, the function
get_and_check_configuration have been renamed into init_mib_parameters
because now, it does init all global variables that represent MIB's
parameters.