handler.c: change the functions' prototype
The functions in hanlder.c that are used as the handlers of the RO-RW
integers, and RO-RO strings did not taken into account the power of the
code that defines a structure "parameter" in  mibParameter.h, this has
been corrected. It was using the name and the value of each parameter
independently when it can use directly the parameter itself.