Maintenance mode is fully implemented
Maintenance mode is now fully implemented: all parameters in the MIB
in the maintenance group support the maintenance mode.
The maintenance mode is tested for RW parameters as follows:
In the case SET_RESERVE_2 of the SNMP requests handler of the object,
we test the value of the globale variable "deviceMode", if it is indeed
maintenanceMode, the code of the handler keeps on running, otherwise it
return "Unvailable ressources" as a response to the request and exit the
handler.