Restore default statement in SET_RESERVE_2 of VFT and CT
The previous commit correct a bug by completiely deleting the default
statement of the switch case performed in SET_RESERVE_2 case of the
handlers of videoFormatTable and channelTable. This works, but it is not
really "clean code", I mean, a switch should have a default statement.
Even an empty one. I added it in this commit (an empty default case for
those switch).