Correct a bug when streaming MPEG-4 videos
There was a data flow error when playing MPEG-4 video even though it was
working fine before. The fix consists to add sync=false to the udpsink
element of the pipeline. The problem was indeed linked to udsink,
launching openvivoe with GST_DEBUG=5 shows that it was this plug-in that
send the error message to the bus. It seems that it was complaining
about the fact that it was not syncronized.