Make openvivoe work with real cameras
Before open VIVOE did work with the webcam only with a MPEG-4 encoding,
but now, it works with all the three encoding we should support. The
trick was to set the pipeline in PAUSE state after creating it. In
stream.c, the function init_stream_SP creates the stream for the
provider. Before it returns, I add a  call to gst_element_set_state,
that set the state of the pipeline to PAUSED. Using it, the pipeline is
just ready to stream and /dev/video is accessed so it does not block
when the call to start_streaming is performed.