Add the initiation of channelUserDesc from conf file
The user already specify the into the configuration file the
"nb-screens" which will define the number of service user. Then for
each service provider, he can defined two parameters, the gst_source
which contains the gstreamer command line to get the vidoe source, and,
now, the channel description he wants to associate to the steam, even if
this parameter is optional.