stream.c: take defaultStartUp mode into account
The VIVOE norm specify that the device can have a defaultStartUp Mode
in which, it can autcomatically start to stream after a reset or a
reboot. This mean that it should not wait to receive the 'start' command
on its channelStatus. This have been implemented in the function
init_stream of stream.c: if the device is indeed in the
DefaultStartMode if so it automatically calls the function
start_streaming.