Provide security on table modification
I notice that SNMP GET requests are handled correctly however, it was
possible to try to modify an object in a table of the MIB with an index
out of range. For example if ethernetIfTable has only 2 indexes, even
though it was possible to perform a SET request on ethernetIfTable.3
(the third index of ethernetIfTable that does not exist of course). So
this lead to a segfault. Indeed, when a value of a table is accessed in
writing, it is saved into a temporary value, so it can be reset if
something goes wrong. But in that case, it tries to copy into that
temporary variable something that does not exist. So here is the
segfault.