Fill channelRtpPt field of SU channel
When a channel that is a serviceUser, receives a steam, it should try to
initiate a maximum of parameter of the MIB, in its channelTable_entry.
Several fields are thus filled. However, it was not the case of
channelRtpPt. Now it is initialized at the same time of the stream.