Add Framerate in channelTable for SU
The channel table's line of service User is filled when the stream is
played through the function fill_entry of the file steam_registration.
However, it did not take into account the video's framerate. Indeed, the
video caps are built when the SDP message is received in the function
get_sdp; This function used the get_media_from_caps() function that will
be released in Gsteamer 1.8 API. But this function do not initialize the
framerate. I had it in this commit.