Initialize channelReceiveIp field for SP
Implicitelly the field channelReceiveIP seems to be more meaningfull for
service user, but the norm does specify that it can be initialize for
service provider too. It will just contained the value of the multicast
Ip on which the channels emit their streams; This have been added into
the code. When the pipeline of service provider are built, especially
the updsink element, the field channelReceivedIp is field by the computed
channel multicast address.