Implement defaultStartUp mode for Service Provider
A Service Provider used in defaultStartUp mode should start to stream
as soon as it has been started up. As each channel, which is a Service
Provider computes its own multicast IP address at the same it is
created, the default multicast IP on which it will launch its stream
should be the same. This has been implemente.d