Implement defaultStartUp mode for SU
DefaultStartUp is the mode in which the device starts streaming if it is
a Service Provider or start to listen for streams, and displays them if
it is a Service User. The mode was implemented for Service Provider but
now it has also been implemented for Service User. However, it implies
several things:
1/ The user needs to register the IP address he wants in the
configuration file.
2/ A service Provider start the stream by calling the function
init_stream() when it receives a valid SAP/SDP message. The problem was
that an entry only listen for SAP/SDP message if it is in the
channelTable_SU list. Moreover, it was added in the list only if a
'start' was received on the channel. The solution is to add it on
startUp, after checking that deviceMode is set to defaultStartUp mode.