Save values entered by user in configuration file
If the user wants to modify the value of defaultReceiveIpAddress of a
service User's channel, he sends a SNMP SET request to modify this value.
But he may want this value to be saved by openvivoe. I add code to
mib-config.c to be able to save the value of defaultReceiveIpAddress
when it is changed by the user.