Save channelUserDesc in conf file when modified
The value of DefaultReceivedIpAddress needs to be save in configuration
when it is modified by the user, but it should be the same for
channelUserDesc. Now it is the case. Furthermore, as the code used to
save the value of DefaultReceivedIpAddress and channelUserDesc is very
similar, I parralellized the work by creating a function in mib_conf
that writes the new value of the key in the configuration file.