Replace xvimagsink by a GstParse from conf file
The sink used by a Service User to display a stream was hardly coded in
pipeline.c. It was building a xvimagesink. This has been replaced by a
command line in the configuration file, so that the user can choose
himself which element to use to sink the received stream.