Start the website with jekyll
GitHub page let us host a website about the project. To do so, GitHub
advises to create a branch name gh-pages to save all web pages created.
It also advices to remove all other files not related to the website on
this branch. This is what I did here.
Furthermore, GitHub is also associated to a static WebPages generator
named Jekyll. I have used it to generate this first set of pages.