Remove index.html from master branch
GitHub provides GitHub Pages to create project's website. The
documentation of GitHub specifies that all the content of those pages
should locates in a branch named gh-pages. I created this branch.
The documentation also specify that the address of our project's website
will be http(s)://openwide-ingenierie.github.io/openvivoe.
However, it seems that having a index.html file in the master branch
overwrite the content of gh-pages, so I removed it. Furthermore as we
will use Jekyll to build our project's pages, this file is useless now.