Get the sink command line before building pipeline
We were gettting the commandline used for the sink only when we were
building the sink for the pipeline of Service User. However, this is not
ideal. Indeed by getting this command line before starting to build the
pipeline, if it is invalid, we would not bother to build the entire
pipeline. This is what is done now.