Move the access to config_file to get sink cmdline to init_stream()
The call to init_sink_from_conf() was performed in create_pipeline_SU()
however we can do it before in init_stream_SU(). This avoids to
perform any calls to functions in pipeline.c if the cmdline enterred by
the user is invalid. It also avoids to include mib-conf.h in pipeline.c.