Create a function stream_SP in stream.c
The function init_stream is used to build the streamings, whatever it is
a Service Provider's streaming or a Service User's steaming. In both
cases, it will call a function init_stream_SP or init_stream_SU. In the
case it calls init_stream_SP, it then test if we are in the default
StartUp mode. If so, it start the steaming. To simplfy all of that, I
create a function named stream_SP that does both steps:
  1/ calls init_stream_SP() to build the pipeline for SP
  2/ check if we are in default start-up mode. If so, then set channel
    Status to start and call channel_status_handler.