Implement vivoe-redirect module
This implement a very new feature in VIVOE: the redirection. This
features is not defined in the Defense Standard 00-82 however it is a
brand new functionnality very usefull. It works as follows:

A device that is defined as both a SP and a SU have at least 2 channels.
One is a videoChannel and is used by the SP part, the other is a Service
User and is used bu the Service Provider part.

On the SU side: the SU is a classical SU. It gets the stream from
another Service Provider (or itself if there is only one device).
However, in the configuration file: vivoe-mibs.conf, it does not use as
gst_sink command line the classical xvimagesink used to display the
stream. Instead it will use a key work "vivoe-redirect". This key work
will seem to be a gsteamer module to the user, but it will not be one.
Indeed this is jsut a KEY WORD. When we build the pipeline for the
Service User, we will read the configuration file to see what is the
command line for gst_sink. If we detect that this command line is set to
"vivoe-redirect" we will immediatly build the a new channel which is a
Service Provider. But it will not have any command line associted to
"gst_source" parameter in configuration file. Instead it will use the
last element of the pipeline of the SU that we described before, and it
will link its pipeline to it. And it will redirect the stream on another
multicast IP address.