Let the user specify the decoder in conf file
Before, on the SU part, after depayloading the stream, we were checking
its format (RAW, MPEG-4, J2K). According to the format detected, we were
decoding the stream. However, this does not let the possibility to the
user to not decode the stream (for a redirection for example). So I
remove the decoding part from the code in pipeline.c. Now the user needs
to specify the decoder direclty in the configuration file.
However, it means that now, each SU channel is specific to a
videoFormat, and it cannot changed the videoFormat it decode.