Change the way to init streamings
Streaming were initialized using the function named init_streaming,
however, this is not convenient. I thought that having one unique
fonction to initialize the streams (for both service providers and
service users) would be a good idea, but actually, it is not. The more
convenient is to have to different functions: init_steam_SP() and
init_streams_SU(). I changed that.

Also I changed the way to initialize the steamings for service
providers. Service Providers are described in the configuration file
under the group [source_x]. So we can nnow how many sources there will
be. Before, we were initializing the number of SP manually in the main.
That is the not the way it should be. Now those streams are initialize
when will call the SNMP deamon and the function
initialize_videoFormatTable_entry (just as for service user).

I also changed the way to initialize videoFormatNumber and
channelNumber. Now we count in the configuration file how many groups
"[source_x]" and "[receiver_y]" there are.