Create a structure redirect_data
To implement the redirection, we need to provide the following mapping:
    SU redirection <---> SP redirection
In other terms which receiver is redirected to which sources. In order
to do that I have created a new structure in "mibParameter.h" that will
hold the videoFormatIndex of the source that will redirect the stream
received by the SU, which will be represented in the structure by its
channelIndex.