Add  redirection channels creation for SP
I have added the code to stream.c that should detect the fact that some
sources are redirections (they are present in the redirect_channels
array). Now if a source is defined as a redirection in the configuration
file openvivoe creates an empty entry in the videoFormatTable and in the
channelTable. It does not set its status to "PLAYING_STATE" even if the
device is set up in "defaultStartUp" mode as this would meant to create
a SDP file to send on a multicast address, but when the source
"redirection" is created it does not have any video so it cannot create
a SDP file.