modify the function channelentry_create_SU_entry()
I changed the name of that function and its parameters. Now it is called
channelentry_create_empty_entry() because, it obvious that for
redirection we have to create empty entries for service providers and
not only service user. Indeed, in the case of a redirection, these sources
(service providers) do not know which stream (video caps) they will use.
It is so impossible to create an entry filled with something, however,
as all the SP are initiated when the SNMP deamon is started, then we
need to create an empty entry in both videoFormatTable and channelTable.
This has been done in stream.c.