Check if SU or SP is a redirection channel
In stream.c it is important to detect if the channel, either it is a
service user or serivce provider channel. Indeed, we will need this
information to build the pipeline: in a case of a service user: it will
change the sink wich will be "intervideosink", in the case of a service
provider it is even more important. Indeed, we need to create a source
element in the pipeline wich will be "intervideosrc" but we cannot create
the rest of the pipeline as we do not know the payloader to add in the
pipeline, as we do not know what kind of video we will redirect.