Implement the redirection functionnality for RAW
The redirection is not implement for RAW videos and works as following:
_ First the sources (service providers are built), if a source is detected
as a "redirection source" we will even though create a pipeline
containing only the "intervideosrc" element, (that we will save into the
sink element for the moment, to be able to retrieve it easily later).
_ Then when a service user receives its SAP/SDP announcement, and is
detected as being a "redirection receiver" it will create a pipeline
that contains a "intervideosink" element.
During the execution of the function init_stream_SU(), after creating
the pipeline for the SU we will call a function created in pipeline.c
called "append_SP_pipeline_for_redirection()" that will append some
element to the pipeline of the SP created in step 1. It will add the RTP
payloader as well as the udpsink (as now we know the caps of the video
we will use). This function append_SP_pipeline_for_redirection() will
also start the SP pipeline in case we are in defaultstartUP mode.

This works well.

However, there still are problems: The first one is that the elements
"intervideosink" and "intervideosrc", only work for RAW video which
means that we maybe will probably have to replace this element by
another that can accept more caps.

The other problem is the case where the user wants to change the
videoFormat of a source that is a redirection. It means that want we
stop a source that is a redirection to change the VF it streams, we will
need to completely delete the pipeline and build a new one just as we
do for SU pipeline.