Implement redirection using shmsink and shmsrc
Redirection was already implemented using intervideosink and
intervideosrc. The main problem was that it could only work with RAW
video. Indeed, the sink and respectively src pads of intervideosrc and
respectively intervideosink only implement RAW video format. A solution
was though to use others gstreamer elements to redirect the video. A
solution I found is the use of the plugin elements shmsrc and shmsink.
These elements write and read into a socket using the specified path in
their parameters "socket-path". The major incovenient of that for now is
that the shmsrc element seems to need a capsfilter after it to specify
the caps of the video it will stream. I have implemented it and it has
worked fine for RAW videos.