Fully implement redirection
Redirection is now fully implemented for RAW, J2K and MPEG-4 videos.
I did not used shmsrc and shmsink as I had thought but apppsink and
appsrc instead. The really main advantage is that appsink/src provide
functions named push-sample that automatically managed caps. No need to
worry about caps anymore. So I have found an example online, and
adapte it to openvivoe, and it work just fine for RAW, MPEG-4 and J2K.
The main modification I did was to add to redirect_data structure 2
GstElement*, one to save the pipeline of the service provider, the other
to save the pipeline of the SU. Indeed, in the functions push-sample
establishes a link between appsrc and appsink. So we need to get the
elements, and the more suitable solution is to retrieve them using their
names from the pipeline. gst_app_sink_pull_sample () allows to retrieve
a sample of data from appsink and gst_app_src_push_sample () to push
this sample to appsrc.