Create log files
I created log.c and log.h that contain function used to display error
messages to the user. For now it is only used for creating gstreamer's
elements and linking them. it allows to have a generic error message
when there is a failure on building a gstreamer's element or on linking
them. It prevent from copying the same code to check if those actions
failed.