Map SU and SP redirection according to their names
This commit provides to the user the possibility to give a name to the
redirection, and that in order to associate the source to the channel.
This is implemented in mib-config.c in init_redirection_data() when we
detect that a gst_source command line is set to "vivoe-redirect
name=toto" we look for the gst_sink commmand line that is also set to
"viveoredirect name=toto". If it is found we register the source's index
x from its group name ( [source_x] ) and the receiver index y form its
group name ([ receiver_y ]) in the redirection data struture. I also
replace the redirect_channels array by a structure redirection that
contains the same array and a "int size" member so I can access easilly to
the data in for loops for example.