Create internal_error global variable
This global variable created is a boolean. It is used to store the fact
that a Gstreamer error have been posted on the bus. Indeed, before if a
Gstreamer's error occured it was displayed on the screen and we were
exiting the main loop. However, during the typefind, the main loop is
indeed running, so when an error occurs, it will be exited, but we would
have kept on executing code. That is not the case anymore, now, we are
properly exiting the program, and we are displaying a error message to the
user.