multicast.c: Set long variable to unsigned long
It does not matter on a 64-bit machine like my x86. But on my raspberry
pi however, on a 32-bit processor, it does matter. Indeed, there was a
problem in the file multicast.c where the  multicast address is
computed. Indeed, I was using long, to compute the decimal value of the
multicast address, on a 64-bit processor, it is stored fine in memory,
as it only take 32 bits. But on a 32-bit processor, the value stored was
negative, so I set the variable to "unsigned long" and it it fine now.