Docs section add "download-mibs"
Actually instaling dependencies was not enough, the user also needs to
download all the MIBs that the openvivoe's MIB is refering to. He can
easily do it by using the package snmp-mibs-downloader. This tools will
automatically download and install the mibs into the correct directory.
I add these explanations into the Docs section of the WebSite under
Header 3 "Try OpenVivoe" and Header 4 "configure-netsnmp".