pipeline.c: Correct a segfault
In pipeline.c we build the sink of SU from the configuration file. If it
is not a redirection we build a bin with no name. However, if this bin
cannot be linked with the rtppayloader added in pipeline ( for example if
the received stream is MPEG-4 but the configuration file gives a file
with no decoder) then we were trying to display an error message. This
error message contained the name of the two elements that could not be
linked, but the bin built did not have a name. I corrected that by adding
a name to the name built from the configuration file.