get_source() remove "get_cmd_line"
In get_source() we were opening the configuration file to get the command
line corresponding to the gst_source. However, this was remaning code
from a previous version of openvivoe. We do not need that code anymore
as a structure redirect_data has been created specifically to do that.
So I remove this from the code.