Docs: Modify the location to place snmpd.conf
In the Docs section, I recomended to copy snmpd.conf to $HOME/.snmp.
However, this was invalid as snmp's deamon does not look for snmpd.conf
in $HOME. So I removed that and ask to the user to copy conf/snmpd.conf
to /etc/snmp and then to restart the snmpd.service deamon so that
changes could take effect.