mib-config.c: improve detection of redirection
Redirections are detected when a source has a gst_source command line set
to:
gst_source=vivoe-redirect name=foo
And as the same time a receiver has a gst_sink command line set to:
gst_sink=vivoe-redirection name=foo

However, if the redirection command line has a gstreamer pipeline
appended to it, the name of the redirection did have a " " (white space)
in it: "name=foo \0" was the actual name, when it should be
"name=foo\0" so there was a problem espacially when there was an
additional pipeline appended to the redirection element.
I change vivoe_redirect() so it returns a name with whitespace.