pipeline.c: Improve redirection
Redirection was implemented but, however, it did not implement the only
feature that make redirection interresting! Indeed openvivoe did parse
the configuration file to detect the redirections defined, but it did
not take into account the remaining pipeline line if there was. As an
example for sink:
gst_sink=videocrop ! vivoe-redirect name=foo
Previously openvivoe did not take into account the element videocrop and
did just build an appsink. Now it takes into account the videocrop, and
link it to the appsink created when "vivoe-redirect" key work is
detected.
I need to implement it for sources now too.