J2K format test is x-j2c, x-jpc or x-j2p
Bedore we were only testing if the JPEG2000 testing format was
image/x-j2c because it was the one contained in the caps obtained after
the opjpegenc. But this opjpegenc as three different formats in src pad:
 x-j2c x-jpc or x-j2p
The first one was selected because it is the first one in the caps of the
sink pad on oppenjpegenc. But all three could be used if another encoder
was also used.