Add the posibility not to filter input video
Before the user had to use only VIVOE video formats and resolutions. But
for testing puposes I needed to be able to use non-VIVOE video formats
and resolutions. So I add a key value to all groups: RAW, MPEG-4 and
JPEG200 which is "no-filter" wich will allow the user to disable VIVOE
filter on its video formats.