stream-conf.c: Remove whitespace from list of keys
It was a problem to have spaces in the key list values of resolution and
encoding for the configuration file vivoe-stream.conf because if there
was a space at the end a the line, the video format was filtered out.
Now the spaces in the list ofkey values will be deleted.