log.c: replace g_printerr by g_warning
The function g_warning is more adapted here as we don't want to exit the
program. An error should completely stop and exit the program. However,
with failures such as "Cannot make element foo" we do not want to exit
the program otherwise all steams connected to this one will be stopped.