Run typefind even if main loop is already running
This commit add the possibility to run typefinds even if the main loop
is already running. To do so the main loop is quit before
performing the typefind. To save the fact that the main loop was running
I created a global boolean "was_running" that is set to TRUE. Then I add
a do/while loop in the main function to run the main loop again if it
has been quit to perform a typefind.