Hope to succeed with redirection
This commit give me hope to succeed to build redirection because I
succeed to run three typefinds in a row on received steam (but only
before the appsink/src elements). These are the pipeline I succeed to
run a typefind on:
udpsrc ! typefind
udpsrc ! queue ! openjepgdec ! typefind
udpsrc ! queue ! openjepgdec ! rtpvrawpay ! typefind

So it seems that I will be at least able to add a fake rtp payloader (As
I wanted to do if I still can't performed typefind after a appsink/src).