pipeline.c:function handle_redirection_SU_pipeline
I add a new function handle_redirection_SU_pipeline use to check the
caps before the appsink element of redirection and add: a capsfilter
with media type imapge/x-jpc if the media type before appsink detected
is JPEG2000 or an element mepg4videoparse if the media type detected
before appsink is MPEG-4. The reasons are:
_For JPEG-2000 that the caps passed to appsink must be linkable with the
rtpj2kpay element (which only accept image/x-jpc media type)
_For MPEG-4 that the typdefind performed in the Service provider's
pipeline will not work is the MPEG 4 parser is not in the same pipeline
as the encoder.