add gst_sink and gst_source to redirect_data str
The structure redirect_data is used to provide a mapping between sources
and receivers for redirection. Furhtermore, rather than accessing
several times to the configuration file what should be more interresting
is to acce it once and save each pipeline in the redirect_data
structure. This is why I added to members in this strucutre: gst_sink
and gst_source to store respectively the gst_source and gst_sink command
line given by the user for the redirection.