Add boolean in stop_data to exit program
There was a little error in the way to exit the program. Indeed it
exists a variable named was_running which is used to run the main loop
again if it has been exited ( normally to perform a typefind ).
However, when SIGINT or SIGTERM were catch we were using the boolean was
running to exit the main function. But One solution is to use the
function exti() with the code EXIT_SUCCESS. The other is the use an
extra boolean in stop_data use to specify that we want quit the
program.