Use main_loop global everywhere
The program was using both the main_loop global variable defined in
mib_parameter.h in some parts of the code and loop given into the
functions' arguments to refer to the GMainLoop of this program. However,
this was not clear. I clarify the code by using the global main_loop
variable so it can be acces in any part of the code.
This lead to changes in particular in detect.c where a completely remove
the structure data_type_detection because it contained as member the
main loop and some GstCaps used to save the detected caps. So I removed
it and now I directly give the caps to the function cb_typefound as
user_date.